app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
mysql-connector-python
